fix(test): safely remove if tmp file exists

+++ pkg/state/state.go
@@ -382,8 +382,10 @@ func Save(state *State) error {
 		return err
 	}
 
-	if err := utils.DeleteFile(tmpFilePath); err != nil {
-		return err
+	if exist := utils.IsFileExists(tmpFilePath); exist {
+		if err := utils.DeleteFile(tmpFilePath); err != nil {
+			return err
+		}
 	}
 
 	return nil
